+++ app/Http/Controllers/HairstylesController.php
@@ -4,30 +4,36 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\hairstyles;
+use Illuminate\Support\Facades\Gate;
+use App\Enums\Permissions;
+use Illuminate\Http\JsonResponse;
 
 class HairstylesController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+
+    public function index($id): JsonResponse
     {
         try {
-            $hairstyles = hairstyles::all();
+            if (Gate::allows(Permissions::ALL_PERMISSION)) {
+                $hairstyles = hairstyles::all();
 
-            if ($hairstyles->isEmpty()) {
-                return response()->json([
-                    "resStatus" => "success",
-                    "message" => "初めまして！新規作成ボタンから使用するヘアスタイルを作成しましょう！",
-                    'hairstyles' => $hairstyles
-                ], 200);
+                if ($hairstyles->isEmpty()) {
+                    return response()->json([
+                        "resStatus" => "success",
+                        "message" => "初めまして！新規作成ボタンから使用するヘアスタイルを作成しましょう！",
+                        'hairstyles' => $hairstyles
+                    ], 200, [], JSON_UNESCAPED_UNICODE)->header('Content-Type', 'application/json; charset=UTF-8');
+                } else {
+                    return response()->json([
+                        "resStatus" => "success",
+                        'hairstyles' => $hairstyles
+                    ], 200, [], JSON_UNESCAPED_UNICODE)->header('Content-Type', 'application/json; charset=UTF-8');
+                }
             } else {
                 return response()->json([
-                    "resStatus" => "success",
-                    'hairstyles' => $hairstyles
-                ], 200);
+                    "resStatus" => "error",
+                    "message" => "権限がありません"
+                ], 500);
             }
         } catch (\Exception $e) {
             return response()->json([
@@ -38,21 +44,28 @@ class HairstylesController extends Controller
     }
 
 
-    public function store(Request $request)
+    public function store(Request $request, $id)
     {
         try {
-            $validatedData = $request->validate([
-                'hairstyle_name' => 'required',
-            ]);
+            if (Gate::allows(Permissions::MANAGER_PERMISSION)) {
+                $validatedData = $request->validate([
+                    'hairstyle_name' => 'required',
+                ]);
 
-            $hairstyle = hairstyles::create([
-                'hairstyle_name' => $validatedData['hairstyle_name'],
-            ]);
+                $hairstyle = hairstyles::create([
+                    'hairstyle_name' => $validatedData['hairstyle_name'],
+                ]);
 
-            return response()->json([
-                "resStatus" => "success",
-                "hairstyle" => $hairstyle
-            ], 200);
+                return response()->json([
+                    "resStatus" => "success",
+                    "hairstyle" => $hairstyle
+                ], 200, [], JSON_UNESCAPED_UNICODE)->header('Content-Type', 'application/json; charset=UTF-8');
+            } else {
+                return response()->json([
+                    "resStatus" => "error",
+                    "message" => "権限がありません"
+                ], 500);
+            }
         } catch (\Exception $e) {
             return response()->json([
                 "resStatus" => "error",
@@ -61,50 +74,58 @@ class HairstylesController extends Controller
         }
     }
 
-    public function show($id)
-    {
-        try {
-            $hairstyle = hairstyles::find($id);
-            if (!$hairstyle) {
-                return response()->json([
-                    'resStatus' => 'error',
-                    'message' => 'ヘアスタイルが見つかりません。'
-                ], 404);
-            }
+    // public function show($id)
+    // {
+    //     try {
 
-            return response()->json([
-                "resStatus" => "success",
-                'hairstyle' => $hairstyle
-            ], 200);
-        } catch (\Exception $e) {
-            return response()->json([
-                "resStatus" => "error",
-                'message' => 'ヘアスタイルが見つかりません。'
-            ], 500);
-        }
-    }
+    //         $hairstyle = hairstyles::find($id);
+    //         if (!$hairstyle) {
+    //             return response()->json([
+    //                 'resStatus' => 'error',
+    //                 'message' => 'ヘアスタイルが見つかりません。'
+    //             ], 404);
+    //         }
+
+    //         return response()->json([
+    //             "resStatus" => "success",
+    //             'hairstyle' => $hairstyle
+    //         ], 200, [], JSON_UNESCAPED_UNICODE)->header('Content-Type', 'application/json; charset=UTF-8');
+    //     } catch (\Exception $e) {
+    //         return response()->json([
+    //             "resStatus" => "error",
+    //             'message' => 'ヘアスタイルが見つかりません。'
+    //         ], 500);
+    //     }
+    // }
 
 
 
     public function update(Request $request, $id)
     {
         try {
-            $validatedData = $request->validate([
-                'hairstyle_name' => 'required',
-            ]);
+            if (Gate::allows(Permissions::MANAGER_PERMISSION)) {
+                $validatedData = $request->validate([
+                    'hairstyle_name' => 'required',
+                ]);
 
-            $hairstyle = hairstyles::find($id);
-            $hairstyle->hairstyle_name = $validatedData['hairstyle_name'];
+                $hairstyle = hairstyles::find($id);
+                $hairstyle->hairstyle_name = $validatedData['hairstyle_name'];
 
-            $hairstyle->save();
+                $hairstyle->save();
 
-            return response()->json(
-                [
-                    "resStatus" => "success",
-                    "hairstyle" => $hairstyle
-                ],
-                200
-            );
+                return response()->json(
+                    [
+                        "resStatus" => "success",
+                        "hairstyle" => $hairstyle
+                    ],
+                    200
+                );
+            } else {
+                return response()->json([
+                    "resStatus" => "error",
+                    "message" => "権限がありません。"
+                ], 500);
+            }
         } catch (\Exception $e) {
             return response()->json([
                 "resStatus" => "error",
@@ -116,18 +137,25 @@ class HairstylesController extends Controller
     public function destroy($id)
     {
         try {
-            $hairstyle = hairstyles::find($id);
-            if (!$hairstyle) {
+            if (Gate::allows(Permissions::OWNER_PERMISSION)) {
+                $hairstyle = hairstyles::find($id);
+                if (!$hairstyle) {
+                    return response()->json([
+                        "resStatus" => "error",
+                        'message' => 'ヘアスタイルが見つかりません。'
+                    ], 500);
+                }
+                $hairstyle->delete();
+                return response()->json([
+                    "resStatus" => "success",
+                    "deleteId" => $id
+                ], 200, [], JSON_UNESCAPED_UNICODE)->header('Content-Type', 'application/json; charset=UTF-8');
+            } else {
                 return response()->json([
                     "resStatus" => "error",
-                    'message' => 'ヘアスタイルが見つかりません。'
+                    "message" => "権限がありません。"
                 ], 500);
             }
-            $hairstyle->delete();
-            return response()->json([
-                "resStatus" => "success",
-                "deleteId" => $id
-            ], 200);
         } catch (\Exception $e) {
             return response()->json([
                 "resStatus" => "error",
